+++ src/app/layout.tsx
@@ -5,6 +5,7 @@ import { type Metadata } from 'next';
 import { Toaster } from 'sonner';
 
 import { TRPCReactProvider } from '@/trpc/react';
+import { ThemeProvider } from '@/components/theme-provider';
 
 const inter = Inter({
   subsets: ['latin'],
@@ -27,13 +28,20 @@ export default function RootLayout({
   children: React.ReactNode;
 }) {
   return (
-    <html lang="en" className={`${inter.variable}`}>
+    <html lang="en" className={`${inter.variable}`} suppressHydrationWarning>
       <body className="font-sans">
-        <TRPCReactProvider>
-          {children}
-          <Analytics />
-          <Toaster richColors position="bottom-right" />
-        </TRPCReactProvider>
+        <ThemeProvider
+          attribute="class"
+          defaultTheme="system"
+          enableSystem
+          disableTransitionOnChange
+        >
+          <TRPCReactProvider>
+            {children}
+            <Analytics />
+            <Toaster richColors position="bottom-right" />
+          </TRPCReactProvider>
+        </ThemeProvider>
       </body>
     </html>
   );